Revert some package name changes to unbreak gesture typing
These classes and methods are expected to have the com.android package
name for swipe/gesture typing to work. Without this change
(or alternatively, a shim on the proprietary library), the keyboard
won't load the proprietary library and will crash when forced to do so.

Note that this is a very, very, very nasty hack. There are more proper
ways to do this. This is only meant as a PoC.