Use correct dictionaries for en_GB and en_AU
With the default main_en dictionary, certain words prefer
the US spelling instead of the region-specific ones due
to their ranking in the wordlist.

To fix this, use a different dictionary that re-ranks the
region-specific spellings and split the English dictionaries
to make sure that the correct dictionary loads.

Both the en_AU and en_GB dictionaries files are taken from LineageOS:
https://github.com/LineageOS/android_vendor_lineage/tree/lineage-18.1/overlay/dictionaries/packages/inputmethods/LatinIME/java/res/raw